+++ src/DbReader/CompositionRoot.cs
@@ -65,7 +65,8 @@
                 .Decorate(typeof(IConstructorSelector), typeof(ConstructorValidator), sr => sr.ImplementingType == typeof(ParameterlessConstructorSelector))
 
                 .Register(factory => DbReaderOptions.ParameterParser, new PerContainerLifetime())
-                .Register(factory => DbReaderOptions.ParameterParser, "ListParameterParser", new PerContainerLifetime());
+                .Register(factory => DbReaderOptions.ListParameterParser, "ListParameterParser", new PerContainerLifetime())
+                .Register<IParameterValidator, ParameterValidator>(new PerContainerLifetime());
         }
 
         private void RegisterScopedServices(IServiceRegistry registry)
